requirements.txt
-e git+https://github.com/mo-ameenx0/AutoNation.git@910d1c2c54f650f247d80aaf0e28ed89854d2b7e#egg=AutoNation
bcrypt==4.1.1
cffi==1.16.0
cryptography==41.0.7
gbulb==0.6.4
paramiko==3.3.1
pycairo==1.25.1
pycparser==2.21
PyGObject==3.46.0
PyNaCl==1.5.0
python-dotenv==1.0.0
